Sprint 4 Fixes, BadRequestException( Class ),  Fixed operations related tests.

+++ src/main/java/dts/logic/operation/RdbOperationsService.java
@@ -14,6 +14,7 @@ import dts.boundaries.OperationBoundary;
 import dts.data.IdGeneratorEntity;
 import dts.data.OperationEntity;
 import dts.logic.IdGeneratorEntityDao;
+import dts.util.BadRequestException;
 
 @Service
 public class RdbOperationsService implements EnhancedOperationsService {
@@ -32,19 +33,25 @@ public class RdbOperationsService implements EnhancedOperationsService {
 	}
 
 	@Override
-	public Object invokeOperation(OperationBoundary operation) {
-		OperationEntity entity = this.operationConverter.toEntity(operation);
+	public Object invokeOperation(OperationBoundary operation) throws Exception {
 
-		IdGeneratorEntity idGeneratorEntity = new IdGeneratorEntity();
-		idGeneratorEntity = this.IdGeneratorEntityDao.save(idGeneratorEntity);
-		UUID newId = idGeneratorEntity.getId();
-		this.IdGeneratorEntityDao.deleteById(newId);
+		if (isValidInput(operation)) {
+			OperationEntity entity = this.operationConverter.toEntity(operation);
 
-		IdBoundary id = new IdBoundary(newId.toString());
-		entity.setOperationId(id.toString());
-		entity.setCreatedTimestamp(new Date());
+			IdGeneratorEntity idGeneratorEntity = new IdGeneratorEntity();
+			idGeneratorEntity = this.IdGeneratorEntityDao.save(idGeneratorEntity);
+			UUID newId = idGeneratorEntity.getId();
+			this.IdGeneratorEntityDao.deleteById(newId);
 
-		return this.operationConverter.toBoundary(this.operationsDao.save(entity));
+			IdBoundary id = new IdBoundary(newId.toString());
+			entity.setOperationId(id.toString());
+			entity.setCreatedTimestamp(new Date());
+
+			return this.operationConverter.toBoundary(this.operationsDao.save(entity));
+		} else {
+			throw new BadRequestException("Cant proccess request due to invalid request message framing");
+		}
+
 	}
 
 	@Override
@@ -65,6 +72,16 @@ public class RdbOperationsService implements EnhancedOperationsService {
 		else
 			throw new Exception("Bad Credentials");
 
+	}
+
+	private boolean isValidInput(OperationBoundary boundary) {
+		boolean valid = true;
+
+		if (boundary.getType() == null || boundary.getType().isEmpty() || boundary.getInvokedBy() == null
+				|| boundary.getItem() == null)
+			valid = false;
+
+		return valid;
 	}
 
 }
